ZON-3306: Move xmlrpc code into its own module
